+++ src/main/java/sysc4806/graduateAdmissions/GraduateAdmissions.java
@@ -7,7 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.annotation.Bean;
 import sysc4806.graduateAdmissions.model.BuddyInfo;
-import sysc4806.graduateAdmissions.model.BuddyInfoRepository;
+import sysc4806.graduateAdmissions.repositories.BuddyInfoRepository;
 
 @SpringBootApplication
 public class GraduateAdmissions {
