+++ src/main/java/sysc4806/graduateAdmissions/repositories/BuddyInfoRepository.java
@@ -1,8 +1,9 @@
-package sysc4806.graduateAdmissions.model;
+package sysc4806.graduateAdmissions.repositories;
 
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
+import sysc4806.graduateAdmissions.model.BuddyInfo;
 
 public interface BuddyInfoRepository extends CrudRepository<BuddyInfo, Long> {
